Refactor how we import rules to make it easier to implement tests
