+++ preciceconfigchecker/cli.py
@@ -6,13 +6,8 @@ import color as c
 
 from precice_config_graph import graph, xml_processing
 
-from rule import check_all_rules, print_all_results
+from rules_processing import check_all_rules, print_all_results
 
-# ALL RULES THAT SHOULD BE CHECKED NEED TO BE IMPORTED
-# SOME IDE's MIGHT REMOVE THEM AS UNUSED IMPORTS
-from rules import missing_coupling
-from rules import missing_exchange
-from rules import data_use_read_write
 
 path: str = None
 debug: bool = False
